refactor: refine code structure
