+++ moat/Cargo.toml
@@ -20,7 +20,6 @@ hmac = { workspace = true }
 pingora = { workspace = true, features = ["proxy", "openssl"] }
 poem = { workspace = true }
 sha2 = { workspace = true }
-tokio = { workspace = true, features = ["rt", "net"] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true, features = ["fmt", "env-filter"] }
 url = { workspace = true }
